factored out some utils from experience with tests
